Rename "TestedNumberCard" to "InspectionsTransitionCard"
