+++ pages/index.vue
@@ -29,7 +29,7 @@
       <patients-breakdown-card />
       <patients-transition-card />
       <patients-detail-card />
-      <tested-number-card />
+      <inspections-transition-card />
       <age-card />
       <sickbeds-used-rate-card />
       <consultations-transition-card />
@@ -45,7 +45,7 @@ import WhatsNew from '@/components/WhatsNew.vue'
 import PatientsBreakdownCard from '@/components/cards/PatientsBreakdownCard.vue'
 import PatientsTransitionCard from '@/components/cards/PatientsTransitionCard.vue'
 import PatientsDetailCard from '@/components/cards/PatientsDetailCard.vue'
-import TestedNumberCard from '@/components/cards/TestedNumberCard.vue'
+import InspectionsTransitionCard from '@/components/cards/InspectionsTransitionCard.vue'
 import AgeCard from '@/components/cards/AgeCard.vue'
 import SickbedsUsedRateCard from '@/components/cards/SickbedsUsedRateCard.vue'
 import ConsultationsTransitionCard from '@/components/cards/ConsultationsTransitionCard.vue'
@@ -57,7 +57,7 @@ export default Vue.extend({
     PatientsBreakdownCard,
     PatientsTransitionCard,
     PatientsDetailCard,
-    TestedNumberCard,
+    InspectionsTransitionCard,
     AgeCard,
     SickbedsUsedRateCard,
     ConsultationsTransitionCard
